wip: Refactor progress
