+++ src/commands/disableUpdatePreview.ts
@@ -1,0 +1,5 @@
+export let disableUpdatePreview = false;
+
+export function updateDisableUpdatePreview(value: boolean) {
+  disableUpdatePreview = value;
+}
