+++ src/logger.ts
@@ -1,0 +1,3 @@
+import * as vscode from 'vscode';
+
+export const logger = vscode.window.createOutputChannel("oil.code", { log: true });
