+++ src/providers/providers.ts
@@ -1,0 +1,5 @@
+import { OilFileSystemProvider } from "./OilFileSystemProvider";
+import { OilPreviewFileSystemProvider } from "./OilPreviewFileSystemProvider";
+
+export const oilFileProvider = new OilFileSystemProvider();
+export const oilPreviewProvider = new OilPreviewFileSystemProvider();
