+++ src/vim/neovim.ts
@@ -1,0 +1,45 @@
+import * as vscode from "vscode";
+import { getDisableVimKeymapsSetting } from "../utils/settings";
+import oilCodeLua from "./oil.code.lua";
+import { logger } from "../logger";
+
+export async function isNeovimAvailable(): Promise<boolean> {
+  try {
+    // Try to execute a simple command provided by the Neovim extension
+    await vscode.commands.executeCommand("vscode-neovim.lua", "return 1");
+    logger.info("Neovim extension is available");
+    return true;
+  } catch (error) {
+    // If command execution fails, the extension is likely not available
+    logger.info("Neovim extension not available or command failed");
+    return false;
+  }
+}
+
+export async function registerNeovimKeymap(): Promise<boolean> {
+  // Check if vim keymaps are disabled in settings
+  const isDisabled = getDisableVimKeymapsSetting();
+  if (isDisabled) {
+    logger.info("Vim keymaps are disabled in settings.");
+    return false;
+  }
+
+  // Check if the extension is available before attempting to register keymaps
+  if (await isNeovimAvailable()) {
+    try {
+      logger.info("Registering Neovim keymaps");
+
+      // Use the Neovim extension's command API to register Lua code
+      await vscode.commands.executeCommand("vscode-neovim.lua", oilCodeLua);
+
+      logger.info("Neovim keymaps registered successfully");
+      return true;
+    } catch (error) {
+      logger.error("Failed to register Neovim keymap:", error);
+      return false;
+    }
+  }
+
+  logger.info("Neovim extension not available, skipping keymap registration");
+  return false;
+}
