+++ src/vim/vimKeymaps.ts
@@ -1,0 +1,47 @@
+import {
+  MAX_EXTENSION_DETECTION_RETRIES,
+  EXTENSION_DETECTION_DELAY,
+} from "../constants";
+import { registerNeovimKeymap } from "./neovim";
+import { registerVSCodeVimKeymap } from "./vscodeVim";
+import { logger } from "../logger";
+
+export async function attemptRegisteringVimKeymaps(
+  retries: number = MAX_EXTENSION_DETECTION_RETRIES,
+  delay: number = EXTENSION_DETECTION_DELAY
+): Promise<void> {
+  let neovimRegistered = false;
+  let vscodevimRegistered = false;
+
+  for (let attempt = 0; attempt <= retries; attempt++) {
+    if (attempt > 0) {
+      logger.info(
+        `Retry attempt ${attempt} of ${retries} to register Vim keymaps`
+      );
+      // Wait before retry
+      await new Promise((resolve) => setTimeout(resolve, delay));
+    }
+
+    // Try to register Neovim keymaps if not already registered
+    if (!neovimRegistered) {
+      neovimRegistered = await registerNeovimKeymap();
+    }
+
+    // Try to register VSCodeVim keymaps if not already registered
+    if (!vscodevimRegistered) {
+      vscodevimRegistered = await registerVSCodeVimKeymap();
+    }
+
+    // If both are registered or we've exhausted attempts, we're done
+    if (neovimRegistered || vscodevimRegistered) {
+      logger.info(
+        "Successfully registered keymaps for all available Vim extensions"
+      );
+      break;
+    }
+  }
+
+  if (!neovimRegistered && !vscodevimRegistered) {
+    logger.info("No Vim extensions were detected after all retry attempts");
+  }
+}
